refactor: align entity structure for settlement
