+++ src/main/java/com/merge/final_project/campaign/campaigns/repository/CampaignRepository.java
@@ -1,5 +1,7 @@
-package com.merge.final_project.campaign.campaigns;
+package com.merge.final_project.campaign.campaigns.repository;
 
+import com.merge.final_project.campaign.campaigns.CampaignStatus;
+import com.merge.final_project.campaign.campaigns.entity.Campaign;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
