+++ src/main/java/com/merge/final_project/report/finalreport/FinalReport.java
@@ -1,7 +1,7 @@
 package com.merge.final_project.report.finalreport;
 
 import com.merge.final_project.auth.useraccount.UsersAccount;
-import com.merge.final_project.campaign.campaigns.Campaign;
+import com.merge.final_project.campaign.campaigns.entity.Campaign;
 import com.merge.final_project.recipient.beneficiary.Beneficiary;
 import jakarta.persistence.*;
 import lombok.Getter;
